Befehl list umgesetzt und umkopieren der Daten bei Beitritt eines neuen Knotens

+++ MyChordDHT/src/Node.java
@@ -26,7 +26,7 @@ public class Node implements Serializable, Hashable {
 		return port;
 	}
 
-	public long getHash() {
+	public int getHash() {
 		return hash;
 	}
